Moving shared variables out into a seperate file
This structure has multiple advantages.

- It is cleaner
- It doesn't require every property to be imported into every Gulp task file
- if a gulp task in one file depends on a gulp task in another file, the dependent file can import the required file

It is mainly just a more correct way to structure the code. It imports the variables it needs from a single file rather than passing them down as props.
