+++ generators/app/templates/gulp/tasks/browserSync.js
@@ -1,23 +1,23 @@
 'use strict';
 
-export default function(gulp, plugins, args, config, taskTarget, browserSync) {
-  // BrowserSync
-  gulp.task('browserSync', () => {
-    return browserSync.init({
-      open: args.open ? 'local' : false,
-      startPath: config.baseUrl,
-      port: config.port || 3000,
-      server: {
-        baseDir: taskTarget,
-        routes: (() => {
-          let routes = {};
+import { gulp, args, config, taskTarget, browserSync } from '../shared-vars';
 
-          // Map base URL to routes
-          routes[config.baseUrl] = taskTarget;
+// BrowserSync
+gulp.task('browserSync', () => {
+  return browserSync.init({
+    open: args.open ? 'local' : false,
+    startPath: config.baseUrl,
+    port: config.port || 3000,
+    server: {
+      baseDir: taskTarget,
+      routes: (() => {
+        let routes = {};
 
-          return routes;
-        })()
-      }
-    });
+        // Map base URL to routes
+        routes[config.baseUrl] = taskTarget;
+
+        return routes;
+      })()
+    }
   });
-}
+});
