+++ generators/app/templates/gulp/tasks/clean.js
@@ -1,13 +1,11 @@
 'use strict';
 
+import { gulp, config, dirs } from '../shared-vars';
+
 import path from 'path';
 import del from 'del';
 
-export default function(gulp, plugins, args, config, taskTarget, browserSync) {
-  let dirs = config.directories;
-
-  // Clean
-  gulp.task('clean', () =>
-    del([path.join(dirs.temporary), path.join(dirs.destination)])
-  );
-}
+// Clean
+gulp.task('clean', () =>
+  del([path.join(dirs.temporary), path.join(dirs.destination)])
+);
