+++ generators/app/templates/gulp/tasks/copy.js
@@ -1,20 +1,19 @@
 'use strict';
 
+import { gulp, plugins, taskTarget, dirs } from '../shared-vars';
+
 import path from 'path';
 
-export default function(gulp, plugins, args, config, taskTarget, browserSync) {
-  let dirs = config.directories;
-  let dest = path.join(taskTarget);
+let dest = path.join(taskTarget);
 
-  // Copy
-  gulp.task('copy', () => {
-    return gulp.src([
-      path.join(dirs.source, '**/*'),
-      '!' + path.join(dirs.source, '{**/\_*,**/\_*/**}')<% if (htmlOption === 'nunjucks') { %>,
-      '!' + path.join(dirs.source, '**/*.nunjucks')<% } else if (htmlOption === 'pug') { %>,
-      '!' + path.join(dirs.source, '**/*.pug')<% } %>
-    ])
-    .pipe(plugins.changed(dest))
-    .pipe(gulp.dest(dest));
-  });
-}
+// Copy
+gulp.task('copy', () => {
+  return gulp.src([
+    path.join(dirs.source, '**/*'),
+    '!' + path.join(dirs.source, '{**/\_*,**/\_*/**}')<% if (htmlOption === 'nunjucks') { %>,
+    '!' + path.join(dirs.source, '**/*.nunjucks')<% } else if (htmlOption === 'pug') { %>,
+    '!' + path.join(dirs.source, '**/*.pug')<% } %>
+  ])
+  .pipe(plugins.changed(dest))
+  .pipe(gulp.dest(dest));
+});
