+++ generators/app/templates/gulp/tasks/eslint.js
@@ -2,33 +2,31 @@
 
 'use strict';
 
+import { gulp, plugins, browserSync, dirs } from '../shared-vars';
+
 import path from 'path';
 import gulpif from 'gulp-if';
 
-export default function(gulp, plugins, args, config, taskTarget, browserSync) {
-  let dirs = config.directories;
-
-  // ESLint
-  gulp.task('eslint', () => {
-    return gulp
-      .src([
-        path.join('gulpfile.babel.js'),
-        path.join(dirs.source, '**/*.js'),
-        // Ignore all vendor folder files
-        '!' + path.join('**/vendor/**', '*')
-      ])
-      .pipe(browserSync.reload({ stream: true, once: true }))
-      .pipe(
-        plugins.eslint({
-          useEslintrc: true
-        })
-      )
-      .pipe(plugins.eslint.format())
-      .pipe(gulpif(!browserSync.active, plugins.eslint.failAfterError()))
-      .on('error', function() {
-        if (!browserSync.active) {
-          process.exit(1);
-        }
-      });
-  });
-}
+// ESLint
+gulp.task('eslint', () => {
+  return gulp
+    .src([
+      path.join('gulpfile.babel.js'),
+      path.join(dirs.source, '**/*.js'),
+      // Ignore all vendor folder files
+      '!' + path.join('**/vendor/**', '*')
+    ])
+    .pipe(browserSync.reload({ stream: true, once: true }))
+    .pipe(
+      plugins.eslint({
+        useEslintrc: true
+      })
+    )
+    .pipe(plugins.eslint.format())
+    .pipe(gulpif(!browserSync.active, plugins.eslint.failAfterError()))
+    .on('error', function() {
+      if (!browserSync.active) {
+        process.exit(1);
+      }
+    });
+});
