+++ generators/app/templates/gulp/tasks/sass.js
@@ -1,45 +1,43 @@
 'use strict';
 
+import { gulp, plugins, args, config, taskTarget, browserSync, dirs, entries } from '../shared-vars';
+
 import path from 'path';
 import autoprefixer from 'autoprefixer';
 import gulpif from 'gulp-if';
 
-export default function(gulp, plugins, args, config, taskTarget, browserSync) {
-  let dirs = config.directories;
-  let entries = config.entries;
-  let dest = path.join(taskTarget, dirs.styles.replace(/^_/, ''));
+let dest = path.join(taskTarget, dirs.styles.replace(/^_/, ''));
 
-  // Sass compilation
-  gulp.task('sass', () => {
-    return gulp
-      .src(path.join(dirs.source, dirs.styles, entries.css))
-      .pipe(plugins.plumber())
-      .pipe(plugins.sourcemaps.init())
-      .pipe(
-        plugins.sass({
-          outputStyle: 'expanded',
-          precision: 10,
-          includePaths: [
-            path.join(dirs.source, dirs.styles),
-            path.join(dirs.source, dirs.modules)
-          ]
-        })
-      )
-      .on('error', function(err) {
-        plugins.util.log(err);
+// Sass compilation
+gulp.task('sass', () => {
+  return gulp
+    .src(path.join(dirs.source, dirs.styles, entries.css))
+    .pipe(plugins.plumber())
+    .pipe(plugins.sourcemaps.init())
+    .pipe(
+      plugins.sass({
+        outputStyle: 'expanded',
+        precision: 10,
+        includePaths: [
+          path.join(dirs.source, dirs.styles),
+          path.join(dirs.source, dirs.modules)
+        ]
       })
-      .on('error', plugins.notify.onError(config.defaultNotification))
-      .pipe(plugins.postcss([autoprefixer()]))
-      .pipe(
-        plugins.rename(function(path) {
-          // Remove 'source' directory as well as prefixed folder underscores
-          // Ex: 'src/_styles' --> '/styles'
-          path.dirname = path.dirname.replace(dirs.source, '').replace('_', '');
-        })
-      )
-      .pipe(gulpif(args.production, plugins.cssnano({ rebase: false })))
-      .pipe(plugins.sourcemaps.write('./'))
-      .pipe(gulp.dest(dest))
-      .pipe(browserSync.stream({ match: '**/*.css' }));
-  });
-}
+    )
+    .on('error', function(err) {
+      plugins.util.log(err);
+    })
+    .on('error', plugins.notify.onError(config.defaultNotification))
+    .pipe(plugins.postcss([autoprefixer()]))
+    .pipe(
+      plugins.rename(function(path) {
+        // Remove 'source' directory as well as prefixed folder underscores
+        // Ex: 'src/_styles' --> '/styles'
+        path.dirname = path.dirname.replace(dirs.source, '').replace('_', '');
+      })
+    )
+    .pipe(gulpif(args.production, plugins.cssnano({ rebase: false })))
+    .pipe(plugins.sourcemaps.write('./'))
+    .pipe(gulp.dest(dest))
+    .pipe(browserSync.stream({ match: '**/*.css' }));
+});
